refactor: move configuration recovery from log into RaftClusterContext
The scan for the newest ConfigurationEntry in the log was private to
ReconfigurationHelper and only reachable from the join path. Recovering
the configuration from the log is cluster-state logic and is about to be
needed from RaftClusterContext.bootstrap() as well, so move it next to
configure()/reset() and let the join path delegate. No behavior change.

(cherry picked from commit 06b47c0d9b9dabaa38ebb67d0e64bbca1238ed35)
